feat: validate inputs; remove: unneccessary logs in server

+++ server/pkg/handlers/stocks_handlers.go
@@ -1,7 +1,6 @@
 package handlers
 
 import (
-	"log"
 	"net/http"
 	"strconv"
 
@@ -36,13 +35,17 @@ func GetStockById(w http.ResponseWriter, r *http.Request) {
 }
 
 func AddStock(w http.ResponseWriter, r *http.Request) {
-	var stock types.Stocks
-	utils.ParseBody(r, &stock)
-	if stock.Stock_id == "" || stock.Prod_id == "" || stock.Expiry_date == "" || stock.Curr_qty == 0 {
+	var s types.Stocks
+	utils.ParseBody(r, &s)
+	if s.Stock_id == "" || s.Prod_id == "" || s.Expiry_date == "" || s.Curr_qty == 0 {
 		responsMessage(w, r, "Error - invalid input json", http.StatusBadRequest, nil)
 		return
 	}
-	err := db.AddStock(stock)
+	if !utils.ValidateStrID(s.Prod_id) || !utils.ValidateStrID(s.Stock_id) || !utils.ValidateDate(s.Expiry_date) {
+		responsMessage(w, r, "Error - invalid input format", http.StatusBadRequest, nil)
+		return
+	}
+	err := db.AddStock(s)
 	if err != nil {
 		responsMessage(w, r, "Error - inserting stocks", http.StatusInternalServerError, err)
 		return
@@ -54,7 +57,10 @@ func RemoveStocks(w http.ResponseWriter, r *http.Request) {
 	vars := mux.Vars(r)
 	prod_id := vars["prod_id"]
 	qty, _ := strconv.Atoi(vars["qty"])
-	log.Println(prod_id)
+    if !utils.ValidateStrID(prod_id) || qty <= 0 {
+		responsMessage(w, r, "Error - invalid input format", http.StatusBadRequest, nil)
+		return
+    }
 	err := db.RemoveStocks(prod_id, qty)
 	if err != nil {
 		responsMessage(w, r, "Error - cannot remove "+strconv.Itoa(qty)+" stocks with product id "+prod_id, http.StatusInternalServerError, err)
